HSLdevcom/bultti#370 Implement new TableReport component to prevent having unnecesary variables.
